pyproject.toml
[project]
name = "git-commenter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gitpython>=3.1.44",
    "google-adk>=0.2.0",
]
